# Skip formatting in the SugaredLogger if the level is disabled