If levels are completely disabled, attempt to avoid the cost of string formatting.