@akshayjshah 2017-01-31
Benchmarks don't cover the formatted logging variants; I'll add some in another commit.