@akshayjshah 2017-01-31
This is good for a whole microsecond 🎉 

Before:
```
BenchmarkZapSugarFmtDisabledLevelsWithoutFields-4        	 1000000	      1196 ns/op	     224 B/op	      12 allocs/op
BenchmarkZapSugarFmtDisabledLevelsAccumulatedContext-4   	 1000000	      1236 ns/op	     224 B/op	      12 allocs/op
BenchmarkZapSugarFmtWithAccumulatedContext-4             	 1000000	      1620 ns/op	     288 B/op	      14 allocs/op
BenchmarkZapSugarFmtWithoutFields-4                      	 1000000	      1655 ns/op	     288 B/op	      14 allocs/op
```

After:
```
BenchmarkZapSugarFmtDisabledLevelsWithoutFields-4        	10000000	       179 ns/op	     104 B/op	       9 allocs/op
BenchmarkZapSugarFmtDisabledLevelsAccumulatedContext-4   	10000000	       184 ns/op	     104 B/op	       9 allocs/op
BenchmarkZapSugarFmtWithAccumulatedContext-4             	 1000000	      1620 ns/op	     288 B/op	      14 allocs/op
BenchmarkZapSugarFmtWithoutFields-4                      	 1000000	      1617 ns/op	     288 B/op	      14 allocs/op
```